chore: move MatchedPattern
