+++ src/Grammar/CollectionPattern.php
@@ -3,7 +3,7 @@
 namespace Phiki\Grammar;
 
 use Phiki\Contracts\PatternCollectionInterface;
-use Phiki\MatchedPattern;
+use Phiki\Grammar\MatchedPattern;
 use Phiki\Tokenizer;
 
 class CollectionPattern extends Pattern implements PatternCollectionInterface
