+++ src/Grammar/IncludePattern.php
@@ -3,7 +3,7 @@
 namespace Phiki\Grammar;
 
 use Phiki\Exceptions\UnrecognisedReferenceException;
-use Phiki\MatchedPattern;
+use Phiki\Grammar\MatchedPattern;
 use Phiki\Tokenizer;
 
 class IncludePattern extends Pattern
