+++ src/Grammar/MatchPattern.php
@@ -3,7 +3,7 @@
 namespace Phiki\Grammar;
 
 use Phiki\Contracts\ContainsCapturesInterface;
-use Phiki\MatchedPattern;
+use Phiki\Grammar\MatchedPattern;
 use Phiki\Support\Regex;
 use Phiki\Tokenizer;
